Add files via upload
tent description:

This file describes what is contained in this repository, which are academic content in JavaScript. To use them, simply load the .html file in your browser along with the file of the same name .js on your machine. Soon, I will put all this content in a repository.